pyproject.toml
[project]
name = "repetit-be-prototype"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Almaas Izdihar", email = "almaasizdihar@gmail.com"},
]
dependencies = [
    "fastapi>=0.112.2",
    "sqlalchemy[asyncio]>=2.0.32",
    "uvicorn>=0.30.6",
    "psycopg[binary]>=3.2.1",
    "alembic>=1.13.2",
    "psycopg2>=2.9.9",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.1",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
